backend/requirements.txt
flask
flask-cors
mysql-connector-python
requests
pyjwt
scikit-learn
pdfminer.six
python-docx
gunicorn
